Reorganized validation files
