+++ src/models/helpers/validation/Validation.ts
@@ -1,0 +1,6 @@
+import {ValidationError} from "./ValidationError";
+
+export interface Validation {
+    warnings: ValidationError[],
+    errors: ValidationError[]
+}
