+++ src/models/helpers/validation/ValidationBase.ts
@@ -1,23 +1,7 @@
-import {ValidationUpdater} from "../helpers/ValidationUpdate";
-export interface Validation {
-    warnings: ValidationError[],
-    errors: ValidationError[]
-}
 
-export interface ValidationError {
-    message: string;
-    loc: string; // property where errors occurred
-}
-
-export interface Validatable {
-    validation: Validation;
-    /**
-     * If object can have children and its validation is composed of child validations,
-     * children will call this method to propagate their new states
-     */
-    setValidationCallback(fn: (err: Validation) => void): void;
-}
-
+import {Validatable} from "./Validatable";
+import {Validation} from "./Validation";
+import {ValidationUpdater} from "./ValidationUpdate";
 
 export abstract class ValidationBase implements Validatable {
