+++ src/models/helpers/validation/ValidationError.ts
@@ -1,0 +1,4 @@
+export interface ValidationError {
+    message: string;
+    loc: string; // property where errors occurred
+}
